requirements.txt
ollama>=0.5.1
trafilatura>=1.12.2
beautifulsoup4>=4.12.3
colorama>=0.4.6
requests>=2.32.4
pyttsx3>=2.99
streamlit>=1.48.0
ddgs>=9.5.2
google-genai>=1.29.0
python-dotenv>=1.1.1